move mostly contents to plugins folder
