+++ .github/ISSUE_TEMPLATE/91011.md
@@ -5,7 +5,7 @@ labels: enhancement
 assignees: OasisAkari
 
 ---
-**请注意，若您想要提出新的功能建议，则您需要自行给出可利用的接口，否则您提出的建议很可能会被关闭。**
+<!--请注意，若您想要提出新的功能建议，则您需要自行给出可利用的接口，否则您提出的建议很可能会被关闭。-->
 **建议描述：**<!--请描述具体问题。-->
 
 **实现方法：**<!-- 在此处填写你认为可以实现的方法。-->
