+++ plugins/ab/__init__.py
@@ -1,11 +1,10 @@
 # -*- coding:utf-8 -*-
 import json
-import re
 
 import aiohttp
 
-from UTC8 import UTC8
-from pbc import pbc2
+from plugins.UTC8 import UTC8
+from plugins.pbc import pbc2
 
 
 async def ab():
