+++ plugins/interwikilist/list.cfg
