+++ plugins/rc/__init__.py
@@ -1,10 +1,9 @@
 # -*- coding:utf-8 -*-
 import aiohttp
 import json
-import re
 
-from UTC8 import UTC8
-from pbc import pbc2
+from plugins.UTC8 import UTC8
+from plugins.pbc import pbc2
 
 
 async def rc():
