+++ plugins/user/ruserlib.py
@@ -3,7 +3,7 @@ import re
 import urllib
 from bs4 import BeautifulSoup as bs
 
-from UTC8 import UTC8
+from plugins.UTC8 import UTC8
 from .tool import yhz, gender
