+++ plugins/user/userlib.py
@@ -1,10 +1,8 @@
 import aiohttp
-import json
-import re
 import re
 import urllib
 
-from UTC8 import UTC8
+from plugins.UTC8 import UTC8
 from .tool import yhz, gender
